move meeting layout to appropriate component
