+++ client/src/components/MeetingsSection/MeetingsList/style.css
@@ -1,3 +1,7 @@
+.meeting-time-details {
+    width: 275px;
+}
+
 .meeting-past {
     background-color: rgba(255, 108, 108, 0.356);
 }
@@ -8,4 +12,8 @@
 
 .meeting-future {
     background-color: rgba(0, 255, 0, 0.61);
+}
+
+button.meeting-student {
+    width: 150px;
 }
